backend/search/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
Django==5.0.1
django-cors-headers==4.3.1
django-elasticsearch-dsl==8.0
django-elasticsearch-dsl-drf==0.22.5
django-environ==0.11.2
django-nine==0.2.7
djangorestframework==3.14.0
drf-yasg==1.21.7
elastic-transport==8.13.0
elasticsearch==8.13.0
elasticsearch-dsl==8.13.0
inflection==0.5.1
packaging==24.0
pika==1.3.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
